+++ fluent_blogs/admin/__init__.py
@@ -1,9 +1,9 @@
 from django.contrib import admin
-from fluent_blogs.admin.entryadmin import AbstractEntryBaseAdmin, EntryAdmin
+from fluent_blogs.admin.entryadmin import AbstractEntryBaseAdmin, EntryAdmin, AbstractEntryBaseAdminForm
 from fluent_blogs.models import get_entry_model, Entry
 
 __all__ = (
-    'AbstractEntryBaseAdmin', 'EntryAdmin',
+    'AbstractEntryBaseAdmin', 'EntryAdmin', 'AbstractEntryBaseAdminForm',
 )