+++ fluent_blogs/models/__init__.py
@@ -1,5 +1,7 @@
 from .db import Entry, AbstractEntry, get_entry_model, get_category_model
+from .managers import EntryManager
 
 __all__ = (
     'Entry', 'AbstractEntry', 'get_entry_model', 'get_category_model'
+    'EntryManager',
 )